requirements.txt
coverage==4.4.1
decorator==4.1.2
Django==1.11.3
django-cors-headers==2.1.0
djangorestframework==3.6.3
drfdocs==0.0.11
Faker==0.7.3
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
Markdown==2.6.8
mixer==5.6.6
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
py==1.4.34
Pygments==2.2.0
pytest==3.1.3
pytest-cov==2.5.1
pytest-django==3.1.2
python-dateutil==2.6.1
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
typing==3.6.1
wcwidth==0.1.7
